requirements.txt
Flask==3.1.0
wakeonlan==3.1.0
mypy==1.15.0
PyYaml==6.0.2
pip==25.0.1
Werkzeug==3.1.3
types-PyYAML==6.0.12.20241230
ruff==0.9.10